requirements.txt
django>=3.1.1,<3.2.0
djangorestframework>=3.12.1,<3.13.0
django-rest-registration>=0.5.6,<0.6.0
flake8>=3.6.0,<3.7.0
ptvsd ==4.3.2
django-cors-headers==3.1.1
randomcolor==0.4.4.6